finished product dump
