+++ be/package.json
@@ -10,8 +10,8 @@
     "dev:hono": "bun run --watch ./index.ts",
     "start": "bun ./index.ts",
     "stop": "node scripts/stop.js",
-    "dump:fe:prods": "bun scripts/fe/products-dump.js",
-    "dump:fe:sitemap": "bun scripts/fe/sitemap-gen.js",
+    "dump:fe:prods": "bun scripts/fe/products-dump",
+    "dump:fe:sitemap": "bun scripts/fe/sitemap-gen",
     "dump:all": "npm run dump:prods && npm run dump:revs",
     "dump:prods": "node scripts/products/dump.js",
     "dump:revs": "node scripts/reviews/dump.js",
